+++ src/Service/Athena/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `us-iso-east-1` region
 - AWS enhancement: Documentation updates.
+- AWS api-change: This release includes support for using Apache Spark in Amazon Athena.
 
 ## 0.1.0